La till en ny Menu.cs
