+++ NameSorter.Upg6.sln
@@ -3,7 +3,12 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.11.35312.102
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NameSorter.Upg6", "NameSorter.Upg6\NameSorter.Upg6.csproj", "{5DCEE3B6-4D71-4A1D-AD60-53937E3573FF}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NameSorter.Upg6", "NameSorter.Upg6\NameSorter.Upg6.csproj", "{5DCEE3B6-4D71-4A1D-AD60-53937E3573FF}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{EA0F43F0-A51B-4DAB-8147-D788124AF95C}"
+	ProjectSection(SolutionItems) = preProject
+		Menu.cs = Menu.cs
+	EndProjectSection
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
